+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Unreleased
 
+### 2.5.5
+- Remove deprecated modules
+- Refine font-weight settings
+- Bump @twreporter/react-components version to 2.1.12
+
 ### 2.5.4
 - [Bug] fix print image abnormally
 - Style change